+++ CHANGELOG.md
@@ -1,1 +1,3 @@
+* 2020/06/29 - Bug fix for Scene Journal and Scene Playlist import refresh.   Correctly repoints those entries to the imported ones.
+* 2020/06/29 - Bug fix for export of thumbname
 * 2020/06/28 - Added localization for English